refactor: move information_schema row mappers into the shared sql package
